Inventory System and Player Stats UI Update
* Big changes to inventory system and player stats system

+++ Assets/Level prototype/z-old/k_enemy.cs
@@ -27,7 +27,7 @@ public class k_enemy : MonoBehaviour
 
     //Attacking
     public float timeBetweenAttacks;
-    public float m_Damage = 20;
+    public int m_Damage = 20;
     bool alreadyAttacked;
